pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "code-assistant"
dynamic = ["version"]
description = "A tool for code analysis and generation using large language models"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Lawrence Godfrey", email = "lawrencegodfrey58@gmail.com"},
]

dependencies = [
    "numpy",
    "fire",
    "openai",
    "python-dotenv",
    "tqdm",
    "torch",
    "transformers",
    "astor",
    "GitPython",
    "einops",
    "numpy",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
]

[project.urls]
Homepage = "https://github.com/Lawrence-Godfrey/code-assist"
Repository = "https://github.com/Lawrence-Godfrey/code-assist.git"
Documentation = "https://github.com/Lawrence-Godfrey/code-assist#readme"

[project.scripts]
code-assistant = "code_assistant.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["code_assistant"]